update: HabitSelection component utillizing new habit context
